modified pets module to integrate with owners

+++ src/pets/pets.service.ts
@@ -4,10 +4,14 @@ import { Repository } from 'typeorm';
 import { CreatePetInput } from './dto/create-pet.input';
 import { UpdatePetInput } from './dto/update-pet.input';
 import { Pet } from './entities/pet.entity';
+import { OwnersService } from '../owners/owners.service';
 
 @Injectable()
 export class PetsService {
-  constructor(@InjectRepository(Pet) private petRepository: Repository<Pet>) {}
+  constructor(
+    @InjectRepository(Pet) private petRepository: Repository<Pet>,
+    private readonly ownersService: OwnersService,
+  ) {}
 
   create(createPetInput: CreatePetInput) {
     const newPet = this.petRepository.create(createPetInput);
@@ -31,6 +35,12 @@ export class PetsService {
 
   async remove(id: number) {
     const deletedPet = await this.findOne(id);
-    return this.petRepository.remove(deletedPet);
+    if (deletedPet) {
+      return this.petRepository.remove(deletedPet);
+    }
+  }
+
+  getOwner(ownerId: number) {
+    return this.ownersService.findOne(ownerId);
   }
 }
